# Fix enums for consumers using isolated modules.