When importing `@lastolivegames/becsy` in a composite project or a project using the `isolatedModules` flag, the following error is encountered:

``` 
node_modules/@lastolivegames/becsy/index.umd.d.ts:104:37 - error TS2748: Cannot access ambient const enums when the '--isolatedModules' flag is provided.

104 type QueryFlavorName = keyof typeof QueryFlavor;
                                        ~~~~~~~~~~~
```

This PR moves all `const enum` to `enum` to avoid the above interoperability issue. 

Let me know what you think!