@pkaminski 2021-09-18
I definitely want to support isolated modules mode, which is why I have `"isolatedModules": true` in `tsconfig.json`.  I don't understand why that doesn't error on the `export const enum`.  Conversely, I don't understand why you get an error for module-local `const enum` declarations, as these are indeed isolated to the module.

While I'd like to understand why the TS compiler isn't working as expected that can often be a foolish quest.  I'm OK with this simple change as long as it doesn't affect performance -- did you try `yarn perf` to see?  If it does, I'd like to find another approach.