@pkaminski 2021-09-18
This [TS issue](https://github.com/microsoft/TypeScript/issues/37774) looks potentially relevant, but it's not (yet) resolved.