@pkaminski 2021-09-18
Thanks!  I agree that performance seems to hold up, though it looks like my efforts to adjust benchmark results according to a dynamic empty loop baseline are completely ineffective.  I'll merge and rerun the benchmarks on my machine to double check.

And yeah, I found that blog post too, but I still can't make sense of what we're seeing.  I do have `isolatedModules` set yet `tsc` isn't warning me about the issues.  I could work around the `QueryFlavor` export by manually constructing the type instead, but I don't want to waste time fighting the language so let's just go with non-`const` enums throughout as the most straightforward solution.