@pkaminski 2021-09-18
Released in 0.8.3 -- thanks again!